+++ svc_trainer.py
@@ -1,12 +1,8 @@
-import os
-import time
-import logging
 import argparse
 import torch
 import torch.multiprocessing as mp
 from omegaconf import OmegaConf
 
-from vits_extend.train import train
 
 torch.backends.cudnn.benchmark = True
 
@@ -19,7 +15,14 @@ if __name__ == '__main__':
                         help="path of checkpoint pt file to resume training")
     parser.add_argument('-n', '--name', type=str, required=True,
                         help="name of the model for logging, saving checkpoint")
+    parser.add_argument('-e', '--extend', type=int, default=0,
+                        help="extend trian")
     args = parser.parse_args()
+
+    if (args.extend > 0):
+        from vits_extend.trainEx import train
+    else:
+        from vits_extend.train import train
 
     hp = OmegaConf.load(args.config)
     with open(args.config, 'r') as f:
